+++ src/TextField/TextField.jsx
@@ -1,6 +1,7 @@
 import FormControl from '@material-ui/core/FormControl'
 import FormHelperText from '@material-ui/core/FormHelperText'
 import InputBase from '@material-ui/core/InputBase'
+import PropTypes from 'prop-types'
 import React from 'react'
 
 import FormLabel from '../form/FormLabel'
@@ -37,10 +38,20 @@ const styles = theme => {
 }
 
 /**
- * The TextField component basically re-exports the Material TextField
- * component, with a few tweaks.
+ * The `<TextField>` component allows users to enter and edit text.
+ *
+ * ```js
+  * import { TextField } from '@theconversation/ui'
+  *
+ * <TextField
+ *   helperText="Enter your full name"
+ *   label="Name"
+ *   onChange={alert("change")}
+ *   placeholder="Jane Doe"
+ * />
+ * ```
  */
-const TextField = ({
+export const TextField = ({
   disabled,
   error,
   fullWidth,
@@ -67,6 +78,68 @@ const TextField = ({
       )}
     </FormControl>
   )
+}
+
+TextField.propTypes = {
+  /**
+   * A boolean value indicating whether the text field is disabled.
+   */
+  disabled: PropTypes.bool,
+
+  /**
+   * A boolean value indicating whether the text field is in an error state.
+   */
+  error: PropTypes.bool,
+
+  /**
+   * A boolean value indicating whether the text field should take up the full
+   * width of its container.
+   */
+  fullWidth: PropTypes.bool,
+
+  /**
+   * The text displayed below the text field.
+   */
+  helperText: PropTypes.string,
+
+  /**
+   * The `id` of the underlying `<input>` element.
+   */
+  id: PropTypes.string,
+
+  /**
+   * The text displayed above the text field.
+   */
+  label: PropTypes.string,
+
+  /**
+   * The callback function called when the text field value changes.
+   */
+  onChange: PropTypes.func,
+
+  /**
+   * The text displayed inside text field before the user enters a value.
+   */
+  placeholder: PropTypes.string,
+
+  /**
+   * A boolean value indicating whether the text field is required to be filled
+   * by the user.
+   */
+  required: PropTypes.bool,
+
+  /**
+   * The type of the text field.
+   */
+  type: PropTypes.oneOf(['text', 'password'])
+}
+
+TextField.defaultProps = {
+  disabled: false,
+  error: false,
+  fullWidth: false,
+  required: false,
+  type: 'text'
 }
 
 export default withStyles(styles)(TextField)